1.4 Update: Support double for divisor config

+++ src/main/java/com/pla/plamoneyget/Config.java
@@ -8,7 +8,7 @@ public class Config {
     public static final ForgeConfigSpec.Builder BUILDER = new ForgeConfigSpec.Builder();
     public static final ForgeConfigSpec SPEC;
 
-    public static ForgeConfigSpec.ConfigValue<Integer> DIVISOR;
+    public static ForgeConfigSpec.ConfigValue<Double> DIVISOR;
     public static ForgeConfigSpec.ConfigValue<Integer> MIN_HEALTH;
     public static ForgeConfigSpec.ConfigValue<List<? extends String>> BLACK_LIST;
     public static ForgeConfigSpec.ConfigValue<List<? extends String>> WHITE_LIST;
@@ -17,7 +17,7 @@ public class Config {
         DIVISOR = BUILDER.comment(
                         "Divisor for money you got from mob.",
                         "Default: 10 => money = mob health / 10")
-                .define("divisor", 10);
+                .define("divisor", 10.0);
         MIN_HEALTH = BUILDER.comment(
                         "Minimum health of a hostile mob that can drop money.",
                         "Default: 20")
